+++ packages/nestjs-util/src/provider/microservice-client/microservice-provider.interface.ts
@@ -28,10 +28,14 @@ export type MicroserviceProviderClientOptions = Omit<RmqOptions['options'], 'url
  * This is the base format which a message queue maps for request responses can be supplied.
  */
 // FIXME: this needs fixing but, but can not type the patterns without typescript going crazy
-export type MicroserviceProviderBaseMessage<Pattern extends Record<string, string>> = Record<
-ValueOf<Pattern>,
-{ request: any | never, response: any | never }
->
+export type MicroserviceProviderBaseMessage<Pattern extends string | symbol | number> = {
+  [K in Pattern]: MicroserviceProviderMessage
+}
+
+export interface MicroserviceProviderMessage {
+  request: any | never
+  response: any | never
+}
 
 /**
  * A timeout exception for message queue internally.
@@ -43,11 +47,27 @@ export class TimeoutException extends RuntimeException {
 }
 
 /**
- * Returns the value of a key from a type map.
+ * Shorthand for enum keys
  */
-export type GetKeyFromTypeMap<P extends any, T extends string | number> = T extends keyof P ? P[T] : never
+export type EnumKeys = string | symbol | number
 
 /**
- * Returns the value of the given key.
+ * Type of a message queue patern thingy
  */
-export type ValueOf<T> = T[keyof T]
+export type BaseMessageQueuePatterns<T extends EnumKeys> = Record<T, EnumKeys>
+
+/**
+ * Type of message queue map
+ */
+export type BaseMessageQueueMap<T extends EnumKeys> = Record<T, MicroserviceProviderBaseMessage<EnumKeys>>
+
+/**
+ * We need this guy for fixing the typescript problem.
+ *
+ * index for not to make typescript go crazy with indexing enums, huge problem
+ *
+ * https://github.com/Microsoft/TypeScript/issues/21987
+ */
+export class BaseMessageIndexes {
+  [k: string]: MicroserviceProviderMessage
+}